Refactored all methods for scoring into a score finder module
